Update.
2002-02-14  Ulrich Drepper  <drepper@redhat.com>

	* stdlib/test-a64l.c (tests): Add more test cases.

	* sunrpc/rtime.c (rtime): Change type of thetime to uint32_t.
	Reported by Walter Harms <WHarms@bfs.de>.

2002-02-11  Jes Sorensen  <jes@trained-monkey.org>

	* sysdeps/unix/sysv/linux/ia64/bits/sigcontext.h: Add sc_flag bit
	definitions for struct sigcontext.
	Suggested by David Mosberger-Tang.

+++ sunrpc/rtime.c
@@ -77,7 +77,8 @@ rtime (struct sockaddr_in *addrp, struct rpc_timeval *timep,
   struct pollfd fd;
   int milliseconds;
   int res;
-  unsigned long thetime;
+  /* RFC 868 says the time is transmitted as a 32-bit value.  */
+  uint32_t thetime;
   struct sockaddr_in from;
   int fromlen;
   int type;
